Moving Mongoid::Geo methods out of Mongoid::Geo::Near
